repo_results/run-20240308-125808_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/url_shortener/13/LLMatic/test_user_accounts.py
import pytest
from user_accounts import UserAccounts

def test_user_accounts():
	user_accounts = UserAccounts()

	# Test registration
	assert user_accounts.register('user1', 'password1') == True
	assert user_accounts.register('user1', 'password1') == False

	# Test login
	assert user_accounts.login('user1', 'password1') == True
	assert user_accounts.login('user1', 'wrongpassword') == False
	assert user_accounts.login('nonexistentuser', 'password1') == False

	# Test URL management
	assert user_accounts.add_url('user1', 'http://example.com') == True
	assert user_accounts.get_urls('user1') == ['http://example.com']
	assert user_accounts.add_url('nonexistentuser', 'http://example.com') == False
	assert user_accounts.get_urls('nonexistentuser') == None
